Menu folders are now deleted when the menu itself is deleted. Signed-off-by:TheDudeFromCI <linktheblademaster@yahoo.com>

+++ src/wraith/lib/util/Algorithms.java
@@ -32,6 +32,9 @@ public class Algorithms{
 		Algorithms.assetFolder = stripQuotes(assetFolder);
 	}
 	public static void deleteFile(File file){
+		if(!file.exists()){
+			return;
+		}
 		if(file.isDirectory()){
 			for(File f : file.listFiles()){
 				deleteFile(f);
